--- yolov5/samples
@@ -1,1 +1,0 @@
-../yolov3-spp/samples/
