+++ yolov5/src/utils.h
@@ -1,34 +1,12 @@
-#ifndef TRTX_YOLOV5_UTILS_H_
-#define TRTX_YOLOV5_UTILS_H_
+#pragma once
 
 #include <dirent.h>
-#include <opencv2/opencv.hpp>
 #include <fstream>
 #include <unordered_map>
 #include <string>
 #include <sstream>
-
-static inline cv::Mat preprocess_img(cv::Mat& img, int input_w, int input_h) {
-    int w, h, x, y;
-    float r_w = input_w / (img.cols * 1.0);
-    float r_h = input_h / (img.rows * 1.0);
-    if (r_h > r_w) {
-        w = input_w;
-        h = r_w * img.rows;
-        x = 0;
-        y = (input_h - h) / 2;
-    } else {
-        w = r_h * img.cols;
-        h = input_h;
-        x = (input_w - w) / 2;
-        y = 0;
-    }
-    cv::Mat re(h, w, CV_8UC3);
-    cv::resize(img, re, re.size(), 0, 0, cv::INTER_LINEAR);
-    cv::Mat out(input_h, input_w, CV_8UC3, cv::Scalar(128, 128, 128));
-    re.copyTo(out(cv::Rect(x, y, re.cols, re.rows)));
-    return out;
-}
+#include <vector>
+#include <cstring>
 
 static inline int read_files_in_dir(const char* p_dir_name, std::vector<std::string>& file_names) {
     DIR *p_dir = opendir(p_dir_name);
@@ -61,6 +39,7 @@ static inline std::string trim_leading_whitespace(const std::string& str) {
     size_t last = str.find_last_not_of(' ');
     return str.substr(first, (last - first + 1));
 }
+
 // Src: https://stackoverflow.com/questions/16605967
 static inline std::string to_string_with_precision(const float a_value, const int n = 2) {
     std::ostringstream out;
@@ -88,6 +67,4 @@ static inline int read_labels(const std::string labels_filename, std::unordered_
 
     return 0;
 }
-
-#endif  // TRTX_YOLOV5_UTILS_H_
