+++ yolov5/yolov5_cls.cpp
@@ -1,333 +1,286 @@
+#include "cuda_utils.h"
+#include "logging.h"
+#include "utils.h"
+#include "model.h"
+#include "config.h"
+
 #include <iostream>
 #include <chrono>
 #include <cmath>
 #include <numeric>
-#include "cuda_utils.h"
-#include "logging.h"
-#include "common.hpp"
-#include "utils.h"
-#include "calibrator.h"
+#include <opencv2/opencv.hpp>
 
-#define USE_FP32  // set USE_INT8 or USE_FP16 or USE_FP32
-#define DEVICE 0  // GPU id
-#define BATCH_SIZE 1
+using namespace nvinfer1;
 
-// stuff we know about the network and the input/output blobs
-static const int INPUT_H = 224;
-static const int INPUT_W = 224;
-static const int CLASS_NUM = 1000;
-
-static const int OUTPUT_SIZE = CLASS_NUM;
-const char* INPUT_BLOB_NAME = "data";
-const char* OUTPUT_BLOB_NAME = "prob";
 static Logger gLogger;
+const static int kOutputSize = kClsNumClass;
 
-static int get_width(int x, float gw, int divisor = 8) {
-    return int(ceil((x * gw) / divisor)) * divisor;
-}
-
-static int get_depth(int x, float gd) {
-    if (x == 1) return 1;
-    int r = round(x * gd);
-    if (x * gd - int(x * gd) == 0.5 && (int(x * gd) % 2) == 0) {
-        --r;
+void batch_preprocess(std::vector<cv::Mat>& imgs, float* output) {
+  for (size_t b = 0; b < imgs.size(); b++) {
+    cv::Mat img;
+    cv::resize(imgs[b], img, cv::Size(kClsInputW, kClsInputH));
+    int i = 0;
+    for (int row = 0; row < img.rows; ++row) {
+      uchar* uc_pixel = img.data + row * img.step;
+      for (int col = 0; col < img.cols; ++col) {
+        output[b * 3 * img.rows * img.cols  + i] = ((float)uc_pixel[2] / 255.0 - 0.485) / 0.229;  // R - 0.485
+        output[b * 3 * img.rows * img.cols + i + img.rows * img.cols] = ((float)uc_pixel[1] / 255.0 - 0.456) / 0.224;
+        output[b * 3 * img.rows * img.cols + i + 2 * img.rows * img.cols] = ((float)uc_pixel[0] / 255.0 - 0.406) / 0.225;
+        uc_pixel += 3;
+        ++i;
+      }
     }
-    return std::max<int>(r, 1);
+  }
 }
 
 std::vector<float> softmax(float *prob, int n) {
-    std::vector<float> res;
-    float sum = 0.0f;
-    float t;
-    for (int i = 0; i < n; i++) {
-        t = expf(prob[i]);
-        res.push_back(t);
-        sum += t;
-    }
-    for (int i = 0; i < n; i++) {
-        res[i] /= sum;
-    }
-    return res;
+  std::vector<float> res;
+  float sum = 0.0f;
+  float t;
+  for (int i = 0; i < n; i++) {
+    t = expf(prob[i]);
+    res.push_back(t);
+    sum += t;
+  }
+  for (int i = 0; i < n; i++) {
+    res[i] /= sum;
+  }
+  return res;
 }
 
 std::vector<int> topk(const std::vector<float>& vec, int k) {
-    std::vector<int> topk_index;
-    std::vector<size_t> vec_index(vec.size());
-    std::iota(vec_index.begin(), vec_index.end(), 0);
+  std::vector<int> topk_index;
+  std::vector<size_t> vec_index(vec.size());
+  std::iota(vec_index.begin(), vec_index.end(), 0);
 
-    std::sort(vec_index.begin(), vec_index.end(), [&vec](size_t index_1, size_t index_2) { return vec[index_1] > vec[index_2]; });
+  std::sort(vec_index.begin(), vec_index.end(), [&vec](size_t index_1, size_t index_2) { return vec[index_1] > vec[index_2]; });
 
-    int k_num = std::min<int>(vec.size(), k);
+  int k_num = std::min<int>(vec.size(), k);
 
-    for (int i = 0; i < k_num; ++i) {
-        topk_index.push_back(vec_index[i]);
-    }
+  for (int i = 0; i < k_num; ++i) {
+    topk_index.push_back(vec_index[i]);
+  }
 
-    return topk_index;
+  return topk_index;
 }
 
 std::vector<std::string> read_classes(std::string file_name) {
-    std::vector<std::string> classes;
-    std::ifstream ifs(file_name, std::ios::in);
-    if (!ifs.is_open()) {
-        std::cerr << file_name << " is not found, pls refer to README and download it." << std::endl;
-        assert(0);
-    }
-    std::string s;
-    while (std::getline(ifs, s)) {
-        classes.push_back(s);
-    }
-    ifs.close();
-    return classes;
-}
-
-ICudaEngine* build_engine(unsigned int maxBatchSize, IBuilder* builder, IBuilderConfig* config, DataType dt, float& gd, float& gw, std::string& wts_name) {
-    INetworkDefinition* network = builder->createNetworkV2(0U);
-
-    // Create input tensor of shape {3, INPUT_H, INPUT_W} with name INPUT_BLOB_NAME
-    ITensor* data = network->addInput(INPUT_BLOB_NAME, dt, Dims3{ 3, INPUT_H, INPUT_W });
-    assert(data);
-    std::map<std::string, Weights> weightMap = loadWeights(wts_name);
-    /* ------ yolov5 backbone------ */
-    auto conv0 = convBlock(network, weightMap, *data,  get_width(64, gw), 6, 2, 1,  "model.0");
-    assert(conv0);
-    auto conv1 = convBlock(network, weightMap, *conv0->getOutput(0), get_width(128, gw), 3, 2, 1, "model.1");
-    auto bottleneck_CSP2 = C3(network, weightMap, *conv1->getOutput(0), get_width(128, gw), get_width(128, gw), get_depth(3, gd), true, 1, 0.5, "model.2");
-    auto conv3 = convBlock(network, weightMap, *bottleneck_CSP2->getOutput(0), get_width(256, gw), 3, 2, 1, "model.3");
-    auto bottleneck_csp4 = C3(network, weightMap, *conv3->getOutput(0), get_width(256, gw), get_width(256, gw), get_depth(6, gd), true, 1, 0.5, "model.4");
-    auto conv5 = convBlock(network, weightMap, *bottleneck_csp4->getOutput(0), get_width(512, gw), 3, 2, 1, "model.5");
-    auto bottleneck_csp6 = C3(network, weightMap, *conv5->getOutput(0), get_width(512, gw), get_width(512, gw), get_depth(9, gd), true, 1, 0.5, "model.6");
-    auto conv7 = convBlock(network, weightMap, *bottleneck_csp6->getOutput(0), get_width(1024, gw), 3, 2, 1, "model.7");
-    auto bottleneck_csp8 = C3(network, weightMap, *conv7->getOutput(0), get_width(1024, gw), get_width(1024, gw), get_depth(3, gd), true, 1, 0.5, "model.8");
-
-    /* ------ yolov5 classification head ------ */
-    auto conv_class = convBlock(network, weightMap, *bottleneck_csp8->getOutput(0), 1280, 1, 1, 1, "model.9.conv");
-    IPoolingLayer* pool2 = network->addPoolingNd(*conv_class->getOutput(0), PoolingType::kAVERAGE, DimsHW{7, 7});
-    assert(pool2);
-    IFullyConnectedLayer* yolo = network->addFullyConnected(*pool2->getOutput(0), CLASS_NUM, weightMap["model.9.linear.weight"], weightMap["model.9.linear.bias"]);
-    assert(yolo);
-
-    yolo->getOutput(0)->setName(OUTPUT_BLOB_NAME);
-    network->markOutput(*yolo->getOutput(0));
-    // Build engine
-    builder->setMaxBatchSize(maxBatchSize);
-    config->setMaxWorkspaceSize(16 * (1 << 20));  // 16MB
-
-#if defined(USE_FP16)
-    config->setFlag(BuilderFlag::kFP16);
-#elif defined(USE_INT8)
-    std::cout << "Your platform support int8: " << (builder->platformHasFastInt8() ? "true" : "false") << std::endl;
-    assert(builder->platformHasFastInt8());
-    config->setFlag(BuilderFlag::kINT8);
-    Int8EntropyCalibrator2* calibrator = new Int8EntropyCalibrator2(1, INPUT_W, INPUT_H, "./coco_calib/", "int8calib.table", INPUT_BLOB_NAME);
-    config->setInt8Calibrator(calibrator);
-#endif
-
-    std::cout << "Building engine, please wait for a while..." << std::endl;
-    ICudaEngine* engine = builder->buildEngineWithConfig(*network, *config);
-    std::cout << "Build engine successfully!" << std::endl;
-
-    // Don't need the network any more
-    network->destroy();
-
-    // Release host memory
-    for (auto& mem : weightMap) {
-        free((void*)(mem.second.values));
-    }
-
-    return engine;
-}
-
-void APIToModel(unsigned int maxBatchSize, IHostMemory** modelStream, float& gd, float& gw, std::string& wts_name) {
-    // Create builder
-    IBuilder* builder = createInferBuilder(gLogger);
-    IBuilderConfig* config = builder->createBuilderConfig();
-
-    // Create model to populate the network, then set the outputs and create an engine
-    ICudaEngine *engine = nullptr;
-
-    engine = build_engine(maxBatchSize, builder, config, DataType::kFLOAT, gd, gw, wts_name);
-    
-    assert(engine != nullptr);
-
-    // Serialize the engine
-    (*modelStream) = engine->serialize();
-
-    // Close everything down
-    engine->destroy();
-    builder->destroy();
-    config->destroy();
-}
-
-void doInference(IExecutionContext& context, cudaStream_t& stream, void **buffers, float* input, float* output, int batchSize) {
-    // infer on the batch asynchronously, and DMA output back to host
-    CUDA_CHECK(cudaMemcpyAsync(buffers[0], input, batchSize * 3 * INPUT_H * INPUT_W * sizeof(float), cudaMemcpyHostToDevice, stream));
-    context.enqueue(batchSize, buffers, stream, nullptr);
-    CUDA_CHECK(cudaMemcpyAsync(output, buffers[1], batchSize * OUTPUT_SIZE * sizeof(float), cudaMemcpyDeviceToHost, stream));
-    cudaStreamSynchronize(stream);
+  std::vector<std::string> classes;
+  std::ifstream ifs(file_name, std::ios::in);
+  if (!ifs.is_open()) {
+    std::cerr << file_name << " is not found, pls refer to README and download it." << std::endl;
+    assert(0);
+  }
+  std::string s;
+  while (std::getline(ifs, s)) {
+    classes.push_back(s);
+  }
+  ifs.close();
+  return classes;
 }
 
 bool parse_args(int argc, char** argv, std::string& wts, std::string& engine, float& gd, float& gw, std::string& img_dir) {
-    if (argc < 4) return false;
-    if (std::string(argv[1]) == "-s" && (argc == 5 || argc == 7)) {
-        wts = std::string(argv[2]);
-        engine = std::string(argv[3]);
-        auto net = std::string(argv[4]);
-        if (net[0] == 'n') {
-            gd = 0.33;
-            gw = 0.25;
-        } else if (net[0] == 's') {
-            gd = 0.33;
-            gw = 0.50;
-        } else if (net[0] == 'm') {
-            gd = 0.67;
-            gw = 0.75;
-        } else if (net[0] == 'l') {
-            gd = 1.0;
-            gw = 1.0;
-        } else if (net[0] == 'x') {
-            gd = 1.33;
-            gw = 1.25;
-        } else if (net[0] == 'c' && argc == 7) {
-            gd = atof(argv[5]);
-            gw = atof(argv[6]);
-        } else {
-            return false;
-        }
-    } else if (std::string(argv[1]) == "-d" && argc == 4) {
-        engine = std::string(argv[2]);
-        img_dir = std::string(argv[3]);
+  if (argc < 4) return false;
+  if (std::string(argv[1]) == "-s" && (argc == 5 || argc == 7)) {
+    wts = std::string(argv[2]);
+    engine = std::string(argv[3]);
+    auto net = std::string(argv[4]);
+    if (net[0] == 'n') {
+      gd = 0.33;
+      gw = 0.25;
+    } else if (net[0] == 's') {
+      gd = 0.33;
+      gw = 0.50;
+    } else if (net[0] == 'm') {
+      gd = 0.67;
+      gw = 0.75;
+    } else if (net[0] == 'l') {
+      gd = 1.0;
+      gw = 1.0;
+    } else if (net[0] == 'x') {
+      gd = 1.33;
+      gw = 1.25;
+    } else if (net[0] == 'c' && argc == 7) {
+      gd = atof(argv[5]);
+      gw = atof(argv[6]);
     } else {
-        return false;
+      return false;
     }
-    return true;
+  } else if (std::string(argv[1]) == "-d" && argc == 4) {
+    engine = std::string(argv[2]);
+    img_dir = std::string(argv[3]);
+  } else {
+    return false;
+  }
+  return true;
+}
+
+void prepare_buffers(ICudaEngine* engine, float** gpu_input_buffer, float** gpu_output_buffer, float** cpu_input_buffer, float** cpu_output_buffer) {
+  assert(engine->getNbBindings() == 2);
+  // In order to bind the buffers, we need to know the names of the input and output tensors.
+  // Note that indices are guaranteed to be less than IEngine::getNbBindings()
+  const int inputIndex = engine->getBindingIndex(kInputTensorName);
+  const int outputIndex = engine->getBindingIndex(kOutputTensorName);
+  assert(inputIndex == 0);
+  assert(outputIndex == 1);
+  // Create GPU buffers on device
+  CUDA_CHECK(cudaMalloc((void**)gpu_input_buffer, kBatchSize * 3 * kClsInputH * kClsInputW * sizeof(float)));
+  CUDA_CHECK(cudaMalloc((void**)gpu_output_buffer, kBatchSize * kOutputSize * sizeof(float)));
+
+  *cpu_input_buffer = new float[kBatchSize * 3 * kClsInputH * kClsInputW];
+  *cpu_output_buffer = new float[kBatchSize * kOutputSize];
+}
+
+void infer(IExecutionContext& context, cudaStream_t& stream, void **buffers, float* input, float* output, int batchSize) {
+  CUDA_CHECK(cudaMemcpyAsync(buffers[0], input, batchSize * 3 * kClsInputH * kClsInputW * sizeof(float), cudaMemcpyHostToDevice, stream));
+  context.enqueue(batchSize, buffers, stream, nullptr);
+  CUDA_CHECK(cudaMemcpyAsync(output, buffers[1], batchSize * kOutputSize * sizeof(float), cudaMemcpyDeviceToHost, stream));
+  cudaStreamSynchronize(stream);
+}
+
+void serialize_engine(unsigned int max_batchsize, float& gd, float& gw, std::string& wts_name, std::string& engine_name) {
+  // Create builder
+  IBuilder* builder = createInferBuilder(gLogger);
+  IBuilderConfig* config = builder->createBuilderConfig();
+
+  // Create model to populate the network, then set the outputs and create an engine
+  ICudaEngine *engine = nullptr;
+
+  engine = build_cls_engine(max_batchsize, builder, config, DataType::kFLOAT, gd, gw, wts_name);
+
+  assert(engine != nullptr);
+
+  // Serialize the engine
+  IHostMemory* serialized_engine = engine->serialize();
+  assert(serialized_engine != nullptr);
+
+  // Save engine to file
+  std::ofstream p(engine_name, std::ios::binary);
+  if (!p) {
+    std::cerr << "Could not open plan output file" << std::endl;
+    assert(false);
+  }
+  p.write(reinterpret_cast<const char*>(serialized_engine->data()), serialized_engine->size());
+
+  // Close everything down
+  engine->destroy();
+  builder->destroy();
+  config->destroy();
+  serialized_engine->destroy();
+}
+
+void deserialize_engine(std::string& engine_name, IRuntime** runtime, ICudaEngine** engine, IExecutionContext** context) {
+  std::ifstream file(engine_name, std::ios::binary);
+  if (!file.good()) {
+    std::cerr << "read " << engine_name << " error!" << std::endl;
+    assert(false);
+  }
+  size_t size = 0;
+  file.seekg(0, file.end);
+  size = file.tellg();
+  file.seekg(0, file.beg);
+  char* serialized_engine = new char[size];
+  assert(serialized_engine);
+  file.read(serialized_engine, size);
+  file.close();
+
+  *runtime = createInferRuntime(gLogger);
+  assert(*runtime);
+  *engine = (*runtime)->deserializeCudaEngine(serialized_engine, size);
+  assert(*engine);
+  *context = (*engine)->createExecutionContext();
+  assert(*context);
+  delete[] serialized_engine;
 }
 
 int main(int argc, char** argv) {
-    cudaSetDevice(DEVICE);
+  cudaSetDevice(kGpuId);
 
-    std::string wts_name = "";
-    std::string engine_name = "";
-    float gd = 0.0f, gw = 0.0f;
-    std::string img_dir;
-    if (!parse_args(argc, argv, wts_name, engine_name, gd, gw, img_dir)) {
-        std::cerr << "arguments not right!" << std::endl;
-        std::cerr << "./yolov5_cls -s [.wts] [.engine] [n/s/m/l/x or c gd gw]  // serialize model to plan file" << std::endl;
-        std::cerr << "./yolov5_cls -d [.engine] ../samples  // deserialize plan file and run inference" << std::endl;
-        return -1;
-    }
+  std::string wts_name = "";
+  std::string engine_name = "";
+  float gd = 0.0f, gw = 0.0f;
+  std::string img_dir;
 
-    // create a model using the API directly and serialize it to a stream
-    if (!wts_name.empty()) {
-        IHostMemory* modelStream{ nullptr };
-        APIToModel(BATCH_SIZE, &modelStream, gd, gw, wts_name);
-        assert(modelStream != nullptr);
-        std::ofstream p(engine_name, std::ios::binary);
-        if (!p) {
-            std::cerr << "could not open plan output file" << std::endl;
-            return -1;
-        }
-        p.write(reinterpret_cast<const char*>(modelStream->data()), modelStream->size());
-        modelStream->destroy();
-        return 0;
-    }
+  if (!parse_args(argc, argv, wts_name, engine_name, gd, gw, img_dir)) {
+    std::cerr << "arguments not right!" << std::endl;
+    std::cerr << "./yolov5_cls -s [.wts] [.engine] [n/s/m/l/x or c gd gw]  // serialize model to plan file" << std::endl;
+    std::cerr << "./yolov5_cls -d [.engine] ../images  // deserialize plan file and run inference" << std::endl;
+    return -1;
+  }
 
-    // deserialize the .engine and run inference
-    std::ifstream file(engine_name, std::ios::binary);
-    if (!file.good()) {
-        std::cerr << "read " << engine_name << " error!" << std::endl;
-        return -1;
-    }
-    char *trtModelStream = nullptr;
-    size_t size = 0;
-    file.seekg(0, file.end);
-    size = file.tellg();
-    file.seekg(0, file.beg);
-    trtModelStream = new char[size];
-    assert(trtModelStream);
-    file.read(trtModelStream, size);
-    file.close();
-
-    std::vector<std::string> file_names;
-    if (read_files_in_dir(img_dir.c_str(), file_names) < 0) {
-        std::cerr << "read_files_in_dir failed." << std::endl;
-        return -1;
-    }
-    auto classes = read_classes("imagenet_classes.txt");
-
-    static float data[BATCH_SIZE * 3 * INPUT_H * INPUT_W];
-    static float prob[BATCH_SIZE * OUTPUT_SIZE];
-    IRuntime* runtime = createInferRuntime(gLogger);
-    assert(runtime != nullptr);
-    ICudaEngine* engine = runtime->deserializeCudaEngine(trtModelStream, size);
-    assert(engine != nullptr);
-    IExecutionContext* context = engine->createExecutionContext();
-    assert(context != nullptr);
-    delete[] trtModelStream;
-    assert(engine->getNbBindings() == 2);
-    void* buffers[2];
-    // In order to bind the buffers, we need to know the names of the input and output tensors.
-    // Note that indices are guaranteed to be less than IEngine::getNbBindings()
-    const int inputIndex = engine->getBindingIndex(INPUT_BLOB_NAME);
-    const int outputIndex = engine->getBindingIndex(OUTPUT_BLOB_NAME);
-    assert(inputIndex == 0);
-    assert(outputIndex == 1);
-    // Create GPU buffers on device
-    CUDA_CHECK(cudaMalloc((void**)&buffers[inputIndex], BATCH_SIZE * 3 * INPUT_H * INPUT_W * sizeof(float)));
-    CUDA_CHECK(cudaMalloc((void**)&buffers[outputIndex], BATCH_SIZE * OUTPUT_SIZE * sizeof(float)));
-
-    // Create stream
-    cudaStream_t stream;
-    CUDA_CHECK(cudaStreamCreate(&stream));
-
-    int fcount = 0;
-    for (int f = 0; f < (int)file_names.size(); f++) {
-        fcount++;
-        if (fcount < BATCH_SIZE && f + 1 != (int)file_names.size()) continue;
-        for (int b = 0; b < fcount; b++) {
-            cv::Mat img = cv::imread(img_dir + "/" + file_names[f - fcount + 1 + b]);
-            if (img.empty()) continue;
-            cv::Mat pr_img;
-            cv::resize(img, pr_img, cv::Size(INPUT_W, INPUT_H));
-            int i = 0;
-            for (int row = 0; row < INPUT_H; ++row) {
-                uchar* uc_pixel = pr_img.data + row * pr_img.step;
-                for (int col = 0; col < INPUT_W; ++col) {
-                    data[b * 3 * INPUT_H * INPUT_W + i] = ((float)uc_pixel[2] / 255.0 - 0.485) / 0.229;  // R - 0.485
-                    data[b * 3 * INPUT_H * INPUT_W + i + INPUT_H * INPUT_W] = ((float)uc_pixel[1] / 255.0 - 0.456) / 0.224;
-                    data[b * 3 * INPUT_H * INPUT_W + i + 2 * INPUT_H * INPUT_W] = ((float)uc_pixel[0] / 255.0 - 0.406) / 0.225;
-                    uc_pixel += 3;
-                    ++i;
-                }
-            }
-        }
-        // Run inference
-        auto start = std::chrono::system_clock::now();
-        doInference(*context, stream, buffers, data, prob, BATCH_SIZE);
-        auto end = std::chrono::system_clock::now();
-        std::cout << "inference time: " << std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count() << "ms" << std::endl;
-        for (int b = 0; b < fcount; b++) {
-            float *p = &prob[b * OUTPUT_SIZE];
-            auto res = softmax(p, OUTPUT_SIZE);
-            auto topk_idx = topk(res, 3);
-            std::cout << file_names[f - fcount + 1 + b] << std::endl;
-            for (auto idx: topk_idx) {
-                std::cout << "  " << classes[idx] << " " << res[idx] << std::endl;
-            }
-        }
-
-        fcount = 0;
-    }
-
-    // Release stream and buffers
-    cudaStreamDestroy(stream);
-    CUDA_CHECK(cudaFree(buffers[inputIndex]));
-    CUDA_CHECK(cudaFree(buffers[outputIndex]));
-    // Destroy the engine
-    context->destroy();
-    engine->destroy();
-    runtime->destroy();
-
+  // Create a model using the API directly and serialize it to a file
+  if (!wts_name.empty()) {
+    serialize_engine(kBatchSize, gd, gw, wts_name, engine_name);
     return 0;
+  }
+
+  // Deserialize the engine from file
+  IRuntime* runtime = nullptr;
+  ICudaEngine* engine = nullptr;
+  IExecutionContext* context = nullptr;
+  deserialize_engine(engine_name, &runtime, &engine, &context);
+  cudaStream_t stream;
+  CUDA_CHECK(cudaStreamCreate(&stream));
+
+  // Prepare cpu and gpu buffers
+  float* gpu_buffers[2];
+  float* cpu_input_buffer = nullptr;
+  float* cpu_output_buffer = nullptr;
+  prepare_buffers(engine, &gpu_buffers[0], &gpu_buffers[1], &cpu_input_buffer, &cpu_output_buffer);
+
+  // Read images from directory
+  std::vector<std::string> file_names;
+  if (read_files_in_dir(img_dir.c_str(), file_names) < 0) {
+    std::cerr << "read_files_in_dir failed." << std::endl;
+    return -1;
+  }
+
+  // Read imagenet labels
+  auto classes = read_classes("imagenet_classes.txt");
+
+  // batch predict
+  for (size_t i = 0; i < file_names.size(); i += kBatchSize) {
+    // Get a batch of images
+    std::vector<cv::Mat> img_batch;
+    std::vector<std::string> img_name_batch;
+    for (size_t j = i; j < i + kBatchSize && j < file_names.size(); j++) {
+      cv::Mat img = cv::imread(img_dir + "/" + file_names[j]);
+      img_batch.push_back(img);
+      img_name_batch.push_back(file_names[j]);
+    }
+
+    // Preprocess
+    batch_preprocess(img_batch, cpu_input_buffer);
+
+    // Run inference
+    auto start = std::chrono::system_clock::now();
+    infer(*context, stream, (void**)gpu_buffers, cpu_input_buffer, cpu_output_buffer, kBatchSize);
+    auto end = std::chrono::system_clock::now();
+    std::cout << "inference time: " << std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count() << "ms" << std::endl;
+
+    // Postprocess and get top-k result
+    for (size_t b = 0; b < img_name_batch.size(); b++) {
+      float* p = &cpu_output_buffer[b * kOutputSize];
+      auto res = softmax(p, kOutputSize);
+      auto topk_idx = topk(res, 3);
+      std::cout << img_name_batch[b] << std::endl;
+      for (auto idx: topk_idx) {
+        std::cout << "  " << classes[idx] << " " << res[idx] << std::endl;
+      }
+    }
+  }
+
+  // Release stream and buffers
+  cudaStreamDestroy(stream);
+  CUDA_CHECK(cudaFree(gpu_buffers[0]));
+  CUDA_CHECK(cudaFree(gpu_buffers[1]));
+  delete[] cpu_input_buffer;
+  delete[] cpu_output_buffer;
+  // Destroy the engine
+  context->destroy();
+  engine->destroy();
+  runtime->destroy();
+
+  return 0;
 }
