+++ yolov5/yolov5_det.cpp
@@ -1,231 +1,18 @@
 #include "cuda_utils.h"
 #include "logging.h"
-#include "common.hpp"
 #include "utils.h"
-#include "calibrator.h"
 #include "preprocess.h"
+#include "postprocess.h"
+#include "model.h"
 
 #include <iostream>
 #include <chrono>
 #include <cmath>
 
-#define USE_FP16  // set USE_INT8 or USE_FP16 or USE_FP32
-#define DEVICE 0  // GPU id
-#define NMS_THRESH 0.4
-#define CONF_THRESH 0.5
-#define MAX_IMAGE_INPUT_SIZE_THRESH 3000 * 3000 // ensure it exceed the maximum size in the input images !
+using namespace nvinfer1;
 
-// stuff we know about the network and the input/output blobs
-static const int kBatchSize = 1;
-static const int kInputH = Yolo::INPUT_H;
-static const int kInputW = Yolo::INPUT_W;
-static const int kNumClass = Yolo::CLASS_NUM;
-static const int kOutputSize = Yolo::MAX_OUTPUT_BBOX_COUNT * sizeof(Yolo::Detection) / sizeof(float) + 1;  // we assume the yololayer outputs no more than MAX_OUTPUT_BBOX_COUNT boxes that conf >= 0.1
-const char* kInputTensorName = "data";
-const char* kOutputTensorName = "prob";
 static Logger gLogger;
-
-static int get_width(int x, float gw, int divisor = 8) {
-    return int(ceil((x * gw) / divisor)) * divisor;
-}
-
-static int get_depth(int x, float gd) {
-    if (x == 1) return 1;
-    int r = round(x * gd);
-    if (x * gd - int(x * gd) == 0.5 && (int(x * gd) % 2) == 0) {
-        --r;
-    }
-    return std::max<int>(r, 1);
-}
-
-static ICudaEngine* build_engine(unsigned int maxBatchSize, IBuilder* builder, IBuilderConfig* config, DataType dt, float& gd, float& gw, std::string& wts_name) {
-    INetworkDefinition* network = builder->createNetworkV2(0U);
-
-    // Create input tensor of shape {3, kInputH, kInputW} with name kInputTensorName
-    ITensor* data = network->addInput(kInputTensorName, dt, Dims3{ 3, kInputH, kInputW });
-    assert(data);
-    std::map<std::string, Weights> weightMap = loadWeights(wts_name);
-    /* ------ yolov5 backbone------ */
-    auto conv0 = convBlock(network, weightMap, *data,  get_width(64, gw), 6, 2, 1,  "model.0");
-    assert(conv0);
-    auto conv1 = convBlock(network, weightMap, *conv0->getOutput(0), get_width(128, gw), 3, 2, 1, "model.1");
-    auto bottleneck_CSP2 = C3(network, weightMap, *conv1->getOutput(0), get_width(128, gw), get_width(128, gw), get_depth(3, gd), true, 1, 0.5, "model.2");
-    auto conv3 = convBlock(network, weightMap, *bottleneck_CSP2->getOutput(0), get_width(256, gw), 3, 2, 1, "model.3");
-    auto bottleneck_csp4 = C3(network, weightMap, *conv3->getOutput(0), get_width(256, gw), get_width(256, gw), get_depth(6, gd), true, 1, 0.5, "model.4");
-    auto conv5 = convBlock(network, weightMap, *bottleneck_csp4->getOutput(0), get_width(512, gw), 3, 2, 1, "model.5");
-    auto bottleneck_csp6 = C3(network, weightMap, *conv5->getOutput(0), get_width(512, gw), get_width(512, gw), get_depth(9, gd), true, 1, 0.5, "model.6");
-    auto conv7 = convBlock(network, weightMap, *bottleneck_csp6->getOutput(0), get_width(1024, gw), 3, 2, 1, "model.7");
-    auto bottleneck_csp8 = C3(network, weightMap, *conv7->getOutput(0), get_width(1024, gw), get_width(1024, gw), get_depth(3, gd), true, 1, 0.5, "model.8");
-    auto spp9 = SPPF(network, weightMap, *bottleneck_csp8->getOutput(0), get_width(1024, gw), get_width(1024, gw), 5, "model.9");
-    /* ------ yolov5 head ------ */
-    auto conv10 = convBlock(network, weightMap, *spp9->getOutput(0), get_width(512, gw), 1, 1, 1, "model.10");
-
-    auto upsample11 = network->addResize(*conv10->getOutput(0));
-    assert(upsample11);
-    upsample11->setResizeMode(ResizeMode::kNEAREST);
-    upsample11->setOutputDimensions(bottleneck_csp6->getOutput(0)->getDimensions());
-
-    ITensor* inputTensors12[] = { upsample11->getOutput(0), bottleneck_csp6->getOutput(0) };
-    auto cat12 = network->addConcatenation(inputTensors12, 2);
-    auto bottleneck_csp13 = C3(network, weightMap, *cat12->getOutput(0), get_width(1024, gw), get_width(512, gw), get_depth(3, gd), false, 1, 0.5, "model.13");
-    auto conv14 = convBlock(network, weightMap, *bottleneck_csp13->getOutput(0), get_width(256, gw), 1, 1, 1, "model.14");
-
-    auto upsample15 = network->addResize(*conv14->getOutput(0));
-    assert(upsample15);
-    upsample15->setResizeMode(ResizeMode::kNEAREST);
-    upsample15->setOutputDimensions(bottleneck_csp4->getOutput(0)->getDimensions());
-
-    ITensor* inputTensors16[] = { upsample15->getOutput(0), bottleneck_csp4->getOutput(0) };
-    auto cat16 = network->addConcatenation(inputTensors16, 2);
-
-    auto bottleneck_csp17 = C3(network, weightMap, *cat16->getOutput(0), get_width(512, gw), get_width(256, gw), get_depth(3, gd), false, 1, 0.5, "model.17");
-
-    /* ------ detect ------ */
-    IConvolutionLayer* det0 = network->addConvolutionNd(*bottleneck_csp17->getOutput(0), 3 * (kNumClass + 5), DimsHW{ 1, 1 }, weightMap["model.24.m.0.weight"], weightMap["model.24.m.0.bias"]);
-    auto conv18 = convBlock(network, weightMap, *bottleneck_csp17->getOutput(0), get_width(256, gw), 3, 2, 1, "model.18");
-    ITensor* inputTensors19[] = { conv18->getOutput(0), conv14->getOutput(0) };
-    auto cat19 = network->addConcatenation(inputTensors19, 2);
-    auto bottleneck_csp20 = C3(network, weightMap, *cat19->getOutput(0), get_width(512, gw), get_width(512, gw), get_depth(3, gd), false, 1, 0.5, "model.20");
-    IConvolutionLayer* det1 = network->addConvolutionNd(*bottleneck_csp20->getOutput(0), 3 * (kNumClass + 5), DimsHW{ 1, 1 }, weightMap["model.24.m.1.weight"], weightMap["model.24.m.1.bias"]);
-    auto conv21 = convBlock(network, weightMap, *bottleneck_csp20->getOutput(0), get_width(512, gw), 3, 2, 1, "model.21");
-    ITensor* inputTensors22[] = { conv21->getOutput(0), conv10->getOutput(0) };
-    auto cat22 = network->addConcatenation(inputTensors22, 2);
-    auto bottleneck_csp23 = C3(network, weightMap, *cat22->getOutput(0), get_width(1024, gw), get_width(1024, gw), get_depth(3, gd), false, 1, 0.5, "model.23");
-    IConvolutionLayer* det2 = network->addConvolutionNd(*bottleneck_csp23->getOutput(0), 3 * (kNumClass + 5), DimsHW{ 1, 1 }, weightMap["model.24.m.2.weight"], weightMap["model.24.m.2.bias"]);
-
-    auto yolo = addYoLoLayer(network, weightMap, "model.24", std::vector<IConvolutionLayer*>{det0, det1, det2});
-    yolo->getOutput(0)->setName(kOutputTensorName);
-    network->markOutput(*yolo->getOutput(0));
-    // Build engine
-    builder->setMaxBatchSize(maxBatchSize);
-    config->setMaxWorkspaceSize(16 * (1 << 20));  // 16MB
-#if defined(USE_FP16)
-    config->setFlag(BuilderFlag::kFP16);
-#elif defined(USE_INT8)
-    std::cout << "Your platform support int8: " << (builder->platformHasFastInt8() ? "true" : "false") << std::endl;
-    assert(builder->platformHasFastInt8());
-    config->setFlag(BuilderFlag::kINT8);
-    Int8EntropyCalibrator2* calibrator = new Int8EntropyCalibrator2(1, kInputW, kInputH, "./coco_calib/", "int8calib.table", kInputTensorName);
-    config->setInt8Calibrator(calibrator);
-#endif
-
-    std::cout << "Building engine, please wait for a while..." << std::endl;
-    ICudaEngine* engine = builder->buildEngineWithConfig(*network, *config);
-    std::cout << "Build engine successfully!" << std::endl;
-
-    // Don't need the network any more
-    network->destroy();
-
-    // Release host memory
-    for (auto& mem : weightMap) {
-        free((void*)(mem.second.values));
-    }
-
-    return engine;
-}
-
-static ICudaEngine* build_engine_p6(unsigned int maxBatchSize, IBuilder* builder, IBuilderConfig* config, DataType dt, float& gd, float& gw, std::string& wts_name) {
-    INetworkDefinition* network = builder->createNetworkV2(0U);
-    // Create input tensor of shape {3, kInputH, kInputW} with name kInputTensorName
-    ITensor* data = network->addInput(kInputTensorName, dt, Dims3{ 3, kInputH, kInputW });
-    assert(data);
-
-    std::map<std::string, Weights> weightMap = loadWeights(wts_name);
-
-    /* ------ yolov5 backbone------ */
-    auto conv0 = convBlock(network, weightMap, *data,  get_width(64, gw), 6, 2, 1,  "model.0");
-    auto conv1 = convBlock(network, weightMap, *conv0->getOutput(0), get_width(128, gw), 3, 2, 1, "model.1");
-    auto c3_2 = C3(network, weightMap, *conv1->getOutput(0), get_width(128, gw), get_width(128, gw), get_depth(3, gd), true, 1, 0.5, "model.2");
-    auto conv3 = convBlock(network, weightMap, *c3_2->getOutput(0), get_width(256, gw), 3, 2, 1, "model.3");
-    auto c3_4 = C3(network, weightMap, *conv3->getOutput(0), get_width(256, gw), get_width(256, gw), get_depth(6, gd), true, 1, 0.5, "model.4");
-    auto conv5 = convBlock(network, weightMap, *c3_4->getOutput(0), get_width(512, gw), 3, 2, 1, "model.5");
-    auto c3_6 = C3(network, weightMap, *conv5->getOutput(0), get_width(512, gw), get_width(512, gw), get_depth(9, gd), true, 1, 0.5, "model.6");
-    auto conv7 = convBlock(network, weightMap, *c3_6->getOutput(0), get_width(768, gw), 3, 2, 1, "model.7");
-    auto c3_8 = C3(network, weightMap, *conv7->getOutput(0), get_width(768, gw), get_width(768, gw), get_depth(3, gd), true, 1, 0.5, "model.8");
-    auto conv9 = convBlock(network, weightMap, *c3_8->getOutput(0), get_width(1024, gw), 3, 2, 1, "model.9");
-    auto c3_10 = C3(network, weightMap, *conv9->getOutput(0), get_width(1024, gw), get_width(1024, gw), get_depth(3, gd), true, 1, 0.5, "model.10");
-    auto sppf11 = SPPF(network, weightMap, *c3_10->getOutput(0), get_width(1024, gw), get_width(1024, gw), 5, "model.11");
-
-    /* ------ yolov5 head ------ */
-    auto conv12 = convBlock(network, weightMap, *sppf11->getOutput(0), get_width(768, gw), 1, 1, 1, "model.12");
-    auto upsample13 = network->addResize(*conv12->getOutput(0));
-    assert(upsample13);
-    upsample13->setResizeMode(ResizeMode::kNEAREST);
-    upsample13->setOutputDimensions(c3_8->getOutput(0)->getDimensions());
-    ITensor* inputTensors14[] = { upsample13->getOutput(0), c3_8->getOutput(0) };
-    auto cat14 = network->addConcatenation(inputTensors14, 2);
-    auto c3_15 = C3(network, weightMap, *cat14->getOutput(0), get_width(1536, gw), get_width(768, gw), get_depth(3, gd), false, 1, 0.5, "model.15");
-
-    auto conv16 = convBlock(network, weightMap, *c3_15->getOutput(0), get_width(512, gw), 1, 1, 1, "model.16");
-    auto upsample17 = network->addResize(*conv16->getOutput(0));
-    assert(upsample17);
-    upsample17->setResizeMode(ResizeMode::kNEAREST);
-    upsample17->setOutputDimensions(c3_6->getOutput(0)->getDimensions());
-    ITensor* inputTensors18[] = { upsample17->getOutput(0), c3_6->getOutput(0) };
-    auto cat18 = network->addConcatenation(inputTensors18, 2);
-    auto c3_19 = C3(network, weightMap, *cat18->getOutput(0), get_width(1024, gw), get_width(512, gw), get_depth(3, gd), false, 1, 0.5, "model.19");
-
-    auto conv20 = convBlock(network, weightMap, *c3_19->getOutput(0), get_width(256, gw), 1, 1, 1, "model.20");
-    auto upsample21 = network->addResize(*conv20->getOutput(0));
-    assert(upsample21);
-    upsample21->setResizeMode(ResizeMode::kNEAREST);
-    upsample21->setOutputDimensions(c3_4->getOutput(0)->getDimensions());
-    ITensor* inputTensors21[] = { upsample21->getOutput(0), c3_4->getOutput(0) };
-    auto cat22 = network->addConcatenation(inputTensors21, 2);
-    auto c3_23 = C3(network, weightMap, *cat22->getOutput(0), get_width(512, gw), get_width(256, gw), get_depth(3, gd), false, 1, 0.5, "model.23");
-
-    auto conv24 = convBlock(network, weightMap, *c3_23->getOutput(0), get_width(256, gw), 3, 2, 1, "model.24");
-    ITensor* inputTensors25[] = { conv24->getOutput(0), conv20->getOutput(0) };
-    auto cat25 = network->addConcatenation(inputTensors25, 2);
-    auto c3_26 = C3(network, weightMap, *cat25->getOutput(0), get_width(1024, gw), get_width(512, gw), get_depth(3, gd), false, 1, 0.5, "model.26");
-
-    auto conv27 = convBlock(network, weightMap, *c3_26->getOutput(0), get_width(512, gw), 3, 2, 1, "model.27");
-    ITensor* inputTensors28[] = { conv27->getOutput(0), conv16->getOutput(0) };
-    auto cat28 = network->addConcatenation(inputTensors28, 2);
-    auto c3_29 = C3(network, weightMap, *cat28->getOutput(0), get_width(1536, gw), get_width(768, gw), get_depth(3, gd), false, 1, 0.5, "model.29");
-
-    auto conv30 = convBlock(network, weightMap, *c3_29->getOutput(0), get_width(768, gw), 3, 2, 1, "model.30");
-    ITensor* inputTensors31[] = { conv30->getOutput(0), conv12->getOutput(0) };
-    auto cat31 = network->addConcatenation(inputTensors31, 2);
-    auto c3_32 = C3(network, weightMap, *cat31->getOutput(0), get_width(2048, gw), get_width(1024, gw), get_depth(3, gd), false, 1, 0.5, "model.32");
-
-    /* ------ detect ------ */
-    IConvolutionLayer* det0 = network->addConvolutionNd(*c3_23->getOutput(0), 3 * (kNumClass + 5), DimsHW{ 1, 1 }, weightMap["model.33.m.0.weight"], weightMap["model.33.m.0.bias"]);
-    IConvolutionLayer* det1 = network->addConvolutionNd(*c3_26->getOutput(0), 3 * (kNumClass + 5), DimsHW{ 1, 1 }, weightMap["model.33.m.1.weight"], weightMap["model.33.m.1.bias"]);
-    IConvolutionLayer* det2 = network->addConvolutionNd(*c3_29->getOutput(0), 3 * (kNumClass + 5), DimsHW{ 1, 1 }, weightMap["model.33.m.2.weight"], weightMap["model.33.m.2.bias"]);
-    IConvolutionLayer* det3 = network->addConvolutionNd(*c3_32->getOutput(0), 3 * (kNumClass + 5), DimsHW{ 1, 1 }, weightMap["model.33.m.3.weight"], weightMap["model.33.m.3.bias"]);
-
-    auto yolo = addYoLoLayer(network, weightMap, "model.33", std::vector<IConvolutionLayer*>{det0, det1, det2, det3});
-    yolo->getOutput(0)->setName(kOutputTensorName);
-    network->markOutput(*yolo->getOutput(0));
-
-    // Build engine
-    builder->setMaxBatchSize(maxBatchSize);
-    config->setMaxWorkspaceSize(16 * (1 << 20));  // 16MB
-#if defined(USE_FP16)
-    config->setFlag(BuilderFlag::kFP16);
-#elif defined(USE_INT8)
-    std::cout << "Your platform support int8: " << (builder->platformHasFastInt8() ? "true" : "false") << std::endl;
-    assert(builder->platformHasFastInt8());
-    config->setFlag(BuilderFlag::kINT8);
-    Int8EntropyCalibrator2* calibrator = new Int8EntropyCalibrator2(1, kInputW, kInputH, "./coco_calib/", "int8calib.table", kInputTensorName);
-    config->setInt8Calibrator(calibrator);
-#endif
-
-    std::cout << "Building engine, please wait for a while..." << std::endl;
-    ICudaEngine* engine = builder->buildEngineWithConfig(*network, *config);
-    std::cout << "Build engine successfully!" << std::endl;
-
-    // Don't need the network any more
-    network->destroy();
-
-    // Release host memory
-    for (auto& mem : weightMap) {
-        free((void*)(mem.second.values));
-    }
-
-    return engine;
-}
+const static int kOutputSize = kMaxNumOutputBbox * sizeof(Detection) / sizeof(float) + 1;
 
 bool parse_args(int argc, char** argv, std::string& wts, std::string& engine, bool& is_p6, float& gd, float& gw, std::string& img_dir) {
   if (argc < 4) return false;
@@ -282,9 +69,9 @@ void prepare_buffers(ICudaEngine* engine, float** gpu_input_buffer, float** gpu_
 }
 
 void infer(IExecutionContext& context, cudaStream_t& stream, void** gpu_buffers, float* output, int batchsize) {
-    context.enqueue(batchsize, gpu_buffers, stream, nullptr);
-    CUDA_CHECK(cudaMemcpyAsync(output, gpu_buffers[1], batchsize * kOutputSize * sizeof(float), cudaMemcpyDeviceToHost, stream));
-    cudaStreamSynchronize(stream);
+  context.enqueue(batchsize, gpu_buffers, stream, nullptr);
+  CUDA_CHECK(cudaMemcpyAsync(output, gpu_buffers[1], batchsize * kOutputSize * sizeof(float), cudaMemcpyDeviceToHost, stream));
+  cudaStreamSynchronize(stream);
 }
 
 void serialize_engine(unsigned int max_batchsize, bool& is_p6, float& gd, float& gw, std::string& wts_name, std::string& engine_name) {
@@ -295,9 +82,9 @@ void serialize_engine(unsigned int max_batchsize, bool& is_p6, float& gd, float&
   // Create model to populate the network, then set the outputs and create an engine
   ICudaEngine *engine = nullptr;
   if (is_p6) {
-    engine = build_engine_p6(max_batchsize, builder, config, DataType::kFLOAT, gd, gw, wts_name);
+    engine = build_det_p6_engine(max_batchsize, builder, config, DataType::kFLOAT, gd, gw, wts_name);
   } else {
-    engine = build_engine(max_batchsize, builder, config, DataType::kFLOAT, gd, gw, wts_name);
+    engine = build_det_engine(max_batchsize, builder, config, DataType::kFLOAT, gd, gw, wts_name);
   }
   assert(engine != nullptr);
 
@@ -345,7 +132,7 @@ void deserialize_engine(std::string& engine_name, IRuntime** runtime, ICudaEngin
 }
 
 int main(int argc, char** argv) {
-  cudaSetDevice(DEVICE);
+  cudaSetDevice(kGpuId);
 
   std::string wts_name = "";
   std::string engine_name = "";
@@ -356,7 +143,7 @@ int main(int argc, char** argv) {
   if (!parse_args(argc, argv, wts_name, engine_name, is_p6, gd, gw, img_dir)) {
     std::cerr << "arguments not right!" << std::endl;
     std::cerr << "./yolov5_det -s [.wts] [.engine] [n/s/m/l/x/n6/s6/m6/l6/x6 or c/c6 gd gw]  // serialize model to plan file" << std::endl;
-    std::cerr << "./yolov5_det -d [.engine] ../samples  // deserialize plan file and run inference" << std::endl;
+    std::cerr << "./yolov5_det -d [.engine] ../images  // deserialize plan file and run inference" << std::endl;
     return -1;
   }
 
@@ -366,7 +153,6 @@ int main(int argc, char** argv) {
     return 0;
   }
 
-
   // Deserialize the engine from file
   IRuntime* runtime = nullptr;
   ICudaEngine* engine = nullptr;
@@ -375,89 +161,72 @@ int main(int argc, char** argv) {
   cudaStream_t stream;
   CUDA_CHECK(cudaStreamCreate(&stream));
 
-  std::vector<std::string> file_names;
-  if (read_files_in_dir(img_dir.c_str(), file_names) < 0) {
-    std::cerr << "read_files_in_dir failed." << std::endl;
-    return -1;
-  }
+  // Init CUDA preprocessing
+  cuda_preprocess_init(kMaxInputImageSize);
 
   // Prepare cpu and gpu buffers
   float* gpu_buffers[2];
   float* cpu_output_buffer = nullptr;
   prepare_buffers(engine, &gpu_buffers[0], &gpu_buffers[1], &cpu_output_buffer);
 
-  uint8_t* img_host = nullptr;
-  uint8_t* img_device = nullptr;
-  // prepare input data cache in pinned memory 
-  CUDA_CHECK(cudaMallocHost((void**)&img_host, MAX_IMAGE_INPUT_SIZE_THRESH * 3));
-  // prepare input data cache in device memory
-  CUDA_CHECK(cudaMalloc((void**)&img_device, MAX_IMAGE_INPUT_SIZE_THRESH * 3));
+  // Read images from directory
+  std::vector<std::string> file_names;
+  if (read_files_in_dir(img_dir.c_str(), file_names) < 0) {
+    std::cerr << "read_files_in_dir failed." << std::endl;
+    return -1;
+  }
 
-  int fcount = 0;
-  std::vector<cv::Mat> imgs_buffer(kBatchSize);
-  for (int f = 0; f < (int)file_names.size(); f++) {
-    fcount++;
-    if (fcount < kBatchSize && f + 1 != (int)file_names.size()) continue;
-    //auto start = std::chrono::system_clock::now();
-    float *buffer_idx = (float*)gpu_buffers[0];
-    for (int b = 0; b < fcount; b++) {
-      cv::Mat img = cv::imread(img_dir + "/" + file_names[f - fcount + 1 + b]);
-      if (img.empty()) continue;
-      imgs_buffer[b] = img;
-      size_t size_image = img.cols * img.rows * 3;
-      size_t size_image_dst = kInputH * kInputW * 3;
-      //copy data to pinned memory
-      memcpy(img_host, img.data, size_image);
-      //copy data to device memory
-      CUDA_CHECK(cudaMemcpyAsync(img_device, img_host, size_image, cudaMemcpyHostToDevice, stream));
-      preprocess_kernel_img(img_device, img.cols, img.rows, buffer_idx, kInputW, kInputH, stream);
-      buffer_idx += size_image_dst;
-      cudaStreamSynchronize(stream);
+  // batch predict
+  for (size_t i = 0; i < file_names.size(); i += kBatchSize) {
+    // Get a batch of images
+    std::vector<cv::Mat> img_batch;
+    std::vector<std::string> img_name_batch;
+    for (size_t j = i; j < i + kBatchSize && j < file_names.size(); j++) {
+      cv::Mat img = cv::imread(img_dir + "/" + file_names[j]);
+      img_batch.push_back(img);
+      img_name_batch.push_back(file_names[j]);
     }
+
+    // Preprocess
+    cuda_batch_preprocess(img_batch, gpu_buffers[0], kInputW, kInputH, stream);
+
     // Run inference
     auto start = std::chrono::system_clock::now();
     infer(*context, stream, (void**)gpu_buffers, cpu_output_buffer, kBatchSize);
     auto end = std::chrono::system_clock::now();
     std::cout << "inference time: " << std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count() << "ms" << std::endl;
-    std::vector<std::vector<Yolo::Detection>> batch_res(fcount);
-    for (int b = 0; b < fcount; b++) {
-      auto& res = batch_res[b];
-      nms(res, &cpu_output_buffer[b * kOutputSize], CONF_THRESH, NMS_THRESH);
+
+    // NMS
+    std::vector<std::vector<Detection>> res_batch;
+    batch_nms(res_batch, cpu_output_buffer, img_batch.size(), kOutputSize, kConfThresh, kNmsThresh);
+
+    // Draw bounding boxes
+    draw_bbox(img_batch, res_batch);
+
+    // Save images
+    for (size_t j = 0; j < img_batch.size(); j++) {
+      cv::imwrite("_" + img_name_batch[j], img_batch[j]);
     }
-    for (int b = 0; b < fcount; b++) {
-      auto& res = batch_res[b];
-      cv::Mat img = imgs_buffer[b];
-      for (size_t j = 0; j < res.size(); j++) {
-        cv::Rect r = get_rect(img, res[j].bbox);
-        cv::rectangle(img, r, cv::Scalar(0x27, 0xC1, 0x36), 2);
-        cv::putText(img, std::to_string((int)res[j].class_id), cv::Point(r.x, r.y - 1), cv::FONT_HERSHEY_PLAIN, 1.2, cv::Scalar(0xFF, 0xFF, 0xFF), 2);
-      }
-      cv::imwrite("_" + file_names[f - fcount + 1 + b], img);
-    }
-    fcount = 0;
   }
 
   // Release stream and buffers
   cudaStreamDestroy(stream);
-  CUDA_CHECK(cudaFree(img_device));
-  CUDA_CHECK(cudaFreeHost(img_host));
   CUDA_CHECK(cudaFree(gpu_buffers[0]));
   CUDA_CHECK(cudaFree(gpu_buffers[1]));
   delete[] cpu_output_buffer;
+  cuda_preprocess_destroy();
   // Destroy the engine
   context->destroy();
   engine->destroy();
   runtime->destroy();
 
-
   // Print histogram of the output distribution
-  //std::cout << "\nOutput:\n\n";
-  //for (unsigned int i = 0; i < kOutputSize; i++)
-  //{
-  //    std::cout << prob[i] << ", ";
-  //    if (i % 10 == 0) std::cout << std::endl;
-  //}
-  //std::cout << std::endl;
+  // std::cout << "\nOutput:\n\n";
+  // for (unsigned int i = 0; i < kOutputSize; i++) {
+  //   std::cout << prob[i] << ", ";
+  //   if (i % 10 == 0) std::cout << std::endl;
+  // }
+  // std::cout << std::endl;
 
   return 0;
 }
